Refactor: update reaction update service
db reaction 상태를 update하는 service의 로직을 refactoring

+++ backend/service/reaction.js
@@ -1,20 +1,21 @@
 import { Reaction } from '../model/Reaction'
 import { verifyRequiredParams, dbErrorHandler } from '../util'
 
-const addReaction = async ({ chatId, workspaceUserInfoId, emoticon }) => {
+const updateReaction = async ({ chatId, workspaceUserInfoId, emoticon }) => {
   verifyRequiredParams(chatId, workspaceUserInfoId, emoticon)
-  const result = await dbErrorHandler(() =>
-    Reaction.create({ chatId, workspaceUserInfoId, emoticon }),
+  const isExist = await dbErrorHandler(() =>
+    Reaction.find({ chatId, workspaceUserInfoId, emoticon }),
   )
-  return result
+  if (isExist.length === 0) {
+    await dbErrorHandler(() =>
+      Reaction.create({ chatId, workspaceUserInfoId, emoticon }),
+    )
+  } else {
+    await dbErrorHandler(() =>
+      Reaction.deleteOne({ chatId, workspaceUserInfoId, emoticon }),
+    )
+  }
+  return isExist.length === 0
 }
 
-const removeReaction = async ({ chatId, workspaceUserInfoId, emoticon }) => {
-  verifyRequiredParams(chatId, workspaceUserInfoId, emoticon)
-  const result = await dbErrorHandler(() =>
-    Reaction.deleteOne({ chatId, workspaceUserInfoId, emoticon }),
-  )
-  return result
-}
-
-module.exports = { addReaction, removeReaction }
+module.exports = { updateReaction }
